Improve stack safety of internal logger functions
This changes the argument for Logger, throwError and triggerExit to be
pass by value so that they are a little safer in a multi-threaded (and
error) environment where the stack and heap change or be corrupted.
These are all inline functions, so the const-ness lets the optimized
assembly pass by value, name or reference.